Move abstractions of entities into their own directory
