+++ src/GameEntityInterfaces/ICard.ts
