+++ src/GameEntityInterfaces/IGame.ts
@@ -1,5 +1,5 @@
 import IRound from './IRound'
-import UniqueIdentifier from '../../../../Utilities/UniqueIdentifier'
+import UniqueIdentifier from '../Utilities/UniqueIdentifier'
 
 interface IGame {
   getCurrentRound(): IRound | null
