+++ src/GameEntityInterfaces/IPlayer.ts
