+++ src/GameEntityInterfaces/IRound.ts
@@ -1,0 +1,8 @@
+import IActionableRound from './IActionableRound'
+import IPlayer from './IPlayer'
+
+interface IRound extends IActionableRound {
+  getCurrentTurnPlayer(): IPlayer | undefined
+}
+
+export default IRound
