+++ src/GameEntityInterfaces/ReadOnlyEntities/IReadOnlyRound.ts
@@ -1,6 +1,6 @@
-import EndOfRoundData from '../Round/EndOfRoundReportData'
+import EndOfRoundData from '../../Entities/Round/EndOfRoundReportData'
 import IReadOnlyTrick from './IReadOnlyTrick'
-import Player from '../Player'
+import Player from '../../Entities/Player'
 
 interface IReadOnlyRound {
   getCurrentTrick(): IReadOnlyTrick
