+++ src/GameEntityInterfaces/ReadOnlyEntities/IReadOnlyTrick.ts
@@ -1,5 +1,5 @@
-import Card from '../Card'
-import TrickData from '../DataStructures/TrickData'
+import Card from '../../Entities/Card'
+import TrickData from '../../Entities/DataStructures/TrickData'
 
 interface IReadOnlyTrick {
   getLeadCard(): Card | undefined
