+++ src/InterfaceAdapters/CommandExecutor/GameCommands/GameCommandFactory.ts
@@ -3,7 +3,7 @@ import BuryCommandDTO from './GameCommandDTOs/BuryCommandDTO'
 import CommandDTO from '../CommandDTO'
 import ICommand from '../ICommand'
 import ICommandFactory from '../ICommandFactory'
-import IGame from './Interfaces/IGame'
+import IGame from '../../../GameEntityInterfaces/IGame'
 import PassCommand from './PassCommand'
 import PickCommand from './PickCommand'
 import PlayAgainCommand from './PlayAgainCommand'
