+++ src/InterfaceAdapters/CommandExecutor/GameCommands/TestClasses/PickCommand.test.ts
@@ -1,6 +1,6 @@
 import ICommand from '../../ICommand'
-import IGame from '../Interfaces/IGame'
-import IRound from '../Interfaces/IRound'
+import IGame from '../../../../GameEntityInterfaces/IGame'
+import IRound from '../../../../GameEntityInterfaces/IRound'
 import PickCommand from '../PickCommand'
 
 describe('Pick Command', () => {
