+++ src/InterfaceAdapters/GamePresenter/GameBoardModel.ts
@@ -1,7 +1,7 @@
 import Card from '../../Entities/Card'
 import EndOfRoundData from '../../Entities/Round/EndOfRoundReportData'
-import IReadOnlyGameModel from '../../Entities/ReadOnlyEntities/IReadOnlyGameModel'
-import IReadOnlyRound from '../../Entities/ReadOnlyEntities/IReadOnlyRound'
+import IReadOnlyGameModel from '../../GameEntityInterfaces/ReadOnlyEntities/IReadOnlyGameModel'
+import IReadOnlyRound from '../../GameEntityInterfaces/ReadOnlyEntities/IReadOnlyRound'
 import ISubscriber from '../../Entities/ISubscriber'
 import Player from '../../Entities/Player'
 import PlayerData from '../../Views/GamePlayViews/EndOfRoundReport/PlayerData'
