+++ src/InterfaceAdapters/GamePresenter/TestClasses/GameBoardModel.test.ts
@@ -2,9 +2,9 @@ import BellePlaineRulesCardRanker from '../../../Entities/BellePlaineRulesCardRa
 import Card from '../../../Entities/Card'
 import CardPlayedByData from '../../../Entities/DataStructures/CardPlayedByData'
 import GameBoardModel from '../GameBoardModel'
-import IReadOnlyGameModel from '../../../Entities/ReadOnlyEntities/IReadOnlyGameModel'
-import IReadOnlyRound from '../../../Entities/ReadOnlyEntities/IReadOnlyRound'
-import IReadOnlyTrick from '../../../Entities/ReadOnlyEntities/IReadOnlyTrick'
+import IReadOnlyGameModel from '../../../GameEntityInterfaces/ReadOnlyEntities/IReadOnlyGameModel'
+import IReadOnlyRound from '../../../GameEntityInterfaces/ReadOnlyEntities/IReadOnlyRound'
+import IReadOnlyTrick from '../../../GameEntityInterfaces/ReadOnlyEntities/IReadOnlyTrick'
 import ISubscriber from '../../../Entities/ISubscriber'
 import Player from '../../../Entities/Player'
 import PlayerLayoutData from '../PlayerLayoutData'
